// src/types/Event.ts
export type EventType = {
  uri: string;
  name: string;
  status: "active" | "canceled";
  start_time: Date;
  end_time: Date;
  event_type: string;
  location:
    | PhysicalLocation
    | OutboundCallLocation
    | InboundCallLocation
    | GoogleConferenceLocation
    | ZoomConferenceLocation
    | GoToMeetingLocation
    | MicrosoftTeamsConferenceLocation
    | WebexConferenceLocation
    | CustomLocation
    | AskInviteeLocation;
  invitees_counter: {
    total: number;
    active: number;
    limit: number;
  };
  created_at: Date;
  updated_at: Date;
  event_memberships: [
    {
      user: string;
    }
  ];
  event_guests: [
    {
      email: string;
      created_at: Date;
      updated_at: Date;
    }
  ];
  cancellation?: Cancellation;
};
// ...
export type PhysicalLocation = {
  type: "physical";
  location: string;
};

export type OutboundCallLocation = {
  type: "outbound_call";
  location: string;
};

export type InboundCallLocation = {
  type: "inbound_call";
  location: string;
};

export type GoogleConferenceLocation = WebConferenceLocation & {
  type: "google_conference";
};

export type ZoomConferenceLocation = WebConferenceLocation & {
  type: "zoom_conference";
  data: {
    id: string;
    password: string;
    settings: {
      global_dial_in_numbers: [global_dial_in_number];
    };
    extra: {
      intl_numbers_url: string;
    };
  };
};

export type GoToMeetingLocation = WebConferenceLocation & {
  type: "gotomeeting";
  data: {
    uniqueMeetingId: string;
    conferenceCallInfo: string;
  };
};

export type MicrosoftTeamsConferenceLocation = WebConferenceLocation & {
  type: "microsoft_teams_conference";
  data: {
    id: string;
    audioConferencing: {
      conferenceId: string;
      dialinUrl: string;
      tollNumber: string;
    };
  };
};

export type WebexConferenceLocation = WebConferenceLocation & {
  type: "webex_conference";
  data: {
    id: string;
    password: string;
    telephony: {
      callInNumbers: [callInNumber];
    };
  };
};

export type CustomLocation = {
  type: "custom";
  location: string;
};

export type AskInviteeLocation = {
  type: "ask_invitee";
  location: string;
};

export type WebConferenceLocation = {
  status: "initiated" | "processing" | "pushed" | "failed";
  join_url: string;
};

export type global_dial_in_number = {
  number: string;
  country: string;
  type: string;
  city: string;
  country_name: string;
};

export type callInNumber = {
  label: string;
  callInNumber: string;
  tollType: string;
};